refactor(core): split update manager progress and release index logic
