+++ docs/architecture/maintainability-baseline.txt
@@ -14,5 +14,5 @@
 625 crates/surge-core/src/releases/restore.rs
 604 crates/surge-core/src/storage/azure.rs
 840 crates/surge-core/src/storage/gcs.rs
-1428 crates/surge-core/src/update/manager.rs
+1169 crates/surge-core/src/update/manager.rs
 760 crates/surge-installer-ui/src/app.rs
